Delegations screen now uses Context API provider.
